+++ src/main/resources/site/parts/subjectArticleList/subjectArticleList.tsx
@@ -4,7 +4,7 @@ import { ChevronDown } from 'react-feather'
 import axios from 'axios'
 import { type SubjectArticleListProps } from '/lib/types/partTypes/subjectArticleList'
 import { type DropdownItem } from '/lib/types/partTypes/publicationArchive'
-import { usePagination } from '/lib/ssb/utils/customHooks/usePagination'
+import { usePagination } from '/lib/ssb/utils/customHooks/paginationHooks'
 
 /* TODO:
 - Fikse sortering (?)
